requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Jinja2==3.1.5
SQLAlchemy==2.0.40
python-dotenv